+++ src/main/java/com/strangegrotto/wealthdraft/networth/projections/Projections.java
@@ -3,11 +3,12 @@ package com.strangegrotto.wealthdraft.networth.projections;
 import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 import org.immutables.value.Value;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Value.Immutable
 public interface Projections {
-    Double getDefaultAnnualGrowth();
+    BigDecimal getDefaultAnnualGrowth();
 
     Map<String, ValOrGerr<ProjectionScenario>> getScenarios();
 }
